refactor(tts): organize tts config components
